Диалог был вынесен в отдельный класс. Добавлены методы для кнопок.
